Encoders: refactor
