+++ tests/data/encoders/test_numpy.py
@@ -5,8 +5,8 @@ from unittest import TestCase
 import numpy as np
 from qiskit.circuit.random import random_circuit
 
-from blackwater.data.encoders.numpy import DefaultNumpyEstimatorInputEncoder
-from blackwater.data.utils import generate_random_pauli_sum_op
+from blackwater.data import DefaultNumpyEstimatorInputEncoder
+from blackwater.data.encoders.utils import generate_random_pauli_sum_op
 
 
 class TestNumpyEncoders(TestCase):
