+++ tests/data/io_tests/test_io.py
@@ -9,7 +9,7 @@ from qiskit.providers.fake_provider import FakeLimaV2
 from qiskit.quantum_info import SparsePauliOp
 
 from blackwater.data.dataio import ExpValDataWriter, ExpValDataReader
-from blackwater.data.encoders.torch import ExpValData
+from blackwater.data.encoders.graph_utils import ExpValData
 from tests.data.encoders.test_graph import create_bell_circuit
