+++ tests/data/loaders/test_dataclasses_loader.py
@@ -10,7 +10,7 @@ from qiskit.quantum_info import SparsePauliOp
 from torch_geometric.data import Data
 
 from blackwater.data.dataio import ExpValDataWriter
-from blackwater.data.encoders.torch import ExpValData
+from blackwater.data.encoders.graph_utils import ExpValData
 from blackwater.data.loaders.dataclasses import ExpValDataSet
 from tests.data.encoders.test_graph import create_bell_circuit
