+++ tests/primitives/test_learning_estimator.py
@@ -10,9 +10,9 @@ from qiskit.providers.fake_provider import FakeLimaV2
 from sklearn.datasets import make_regression
 from sklearn.ensemble import RandomForestRegressor
 
-from blackwater.data.encoders.numpy import DefaultNumpyEstimatorInputEncoder
-from blackwater.data.encoders.torch import DefaultPyGEstimatorEncoder
-from blackwater.data.utils import generate_random_pauli_sum_op
+from blackwater.data import DefaultNumpyEstimatorInputEncoder
+from blackwater.data.encoders.graph_utils import DefaultPyGEstimatorEncoder
+from blackwater.data.encoders.utils import generate_random_pauli_sum_op
 from blackwater.primitives.learning_esimator import (
     learning_estimator,
     ScikitLearnEstimatorModel,
